fix(02-11): clear the seven clippy lints blocking a -D warnings gate
- tools.rs: point-free map_err; module-scoped allow for the enum variant
  postfix, since tool_box! generates the variant names and an attribute on
  the invocation is ignored
- app.rs: drop four f32 -> f32 casts in the two viewport constructions
- tabs.rs: accept TabManager::open's arity, one caller, all engine handles

Behaviour unchanged: cargo test 9/9 and the full e2e suite 14/14 PASS,
including takeover_test and keyboard_nav_test which drive the mouse paths
the cast removal touched.

+++ crates/talaria-shell/src/app.rs
@@ -988,7 +988,7 @@ fn forward_mouse_move(state: &Shared, position: PhysicalPosition<f64>) {
     let size = webview.size();
     let viewport = euclid::Rect::new(
         euclid::Point2D::zero(),
-        euclid::Size2D::new(size.width as f32, size.height as f32),
+        euclid::Size2D::new(size.width, size.height),
     );
     if !viewport.contains(point) {
         if viewport.contains(previous) {
@@ -1013,7 +1013,7 @@ fn forward_mouse_button(
     let size = webview.size();
     let viewport = euclid::Rect::new(
         euclid::Point2D::zero(),
-        euclid::Size2D::new(size.width as f32, size.height as f32),
+        euclid::Size2D::new(size.width, size.height),
     );
     if !viewport.contains(point) {
         return;
